requirements.txt
fastapi==0.112.2
uvicorn==0.24.0
pydantic==2.12.0
sqlmodel==0.0.18
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
supabase==2.4.3
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
python-multipart==0.0.9
requests==2.31.0
bcrypt==5.0.0
pydantic-settings==2.13.1